backend/config.py
"""
Backend Configuration — Churn Prediction Microservice.

Loads environment variables and sets defaults.
"""

import os
from dotenv import load_dotenv

load_dotenv()

# Supabase (future integration)
SUPABASE_URL = os.getenv("SUPABASE_URL", "")
SUPABASE_KEY = os.getenv("SUPABASE_KEY", "")

# Server
HOST = os.getenv("API_HOST", "0.0.0.0")
PORT = int(os.getenv("API_PORT", "8000"))

# Model
MODEL_DIR = os.path.join(os.path.dirname(os.path.abspath(__file__)), "models")
DATA_DIR = os.path.join(os.path.dirname(os.path.abspath(__file__)), "data")
