backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pandas==2.2.3
scikit-learn==1.5.2
numpy==1.26.4
pydantic==2.9.0
python-dotenv==1.0.1
